# Revert changes that don't improve performance